+++ include/BDSEventAction.hh
@@ -64,6 +64,16 @@ private:
   G4double starts; ///< Precise start time in seconds.
   G4double stops;  ///< Precise stop time in seconds.
 
+  /// @{ Cache of variable from global constants.
+  G4double trajectoryEnergyThreshold;
+  G4double trajectoryCutZ;
+  G4double trajectoryCutR;
+  G4bool   trajConnect;
+  G4String particleToStore;
+  G4String particleIDToStore;
+  G4int    depth;
+  /// @}
+
   std::string seedStateAtStart; ///< Seed state at start of the event.
 
   /// A copy of the pointer to event info instance that is registered to the event. Geant4