+++ parser/field.cc
@@ -51,6 +51,7 @@ void Field::clear()
   angle     = 0;
   axisAngle = false;
   autoScale = false;
+  maximumStepLength = -1;
 }
 
 void Field::PublishMembers()
@@ -78,6 +79,7 @@ void Field::PublishMembers()
   publish("angle",                &Field::angle);
   publish("axisAngle",            &Field::axisAngle);
   publish("autoScale",            &Field::autoScale);
+  publish("maximumStepLength",    &Field::maximumStepLength);
 }
 
 void Field::print()const
@@ -97,5 +99,6 @@ void Field::print()const
 	    << "axisX, Y, Z "          << axisX << " " << axisY << " " << axisZ << std::endl
 	    << "angle "                << angle                << std::endl
 	    << "axisAngle "            << axisAngle            << std::endl
-	    << "autoScale "            << autoScale            << std::endl;
+	    << "autoScale "            << autoScale            << std::endl
+	    << "maximumStepLength "    << maximumStepLength    << std::endl;
 }